requirements.txt
Django
djangorestframework
gunicorn
whitenoise
dj-database-url
python-dotenv
psycopg2-binary
requests
